Rename 3 component "base" modules for consistency. (#440)
