+++ packages/top-app-bar-fixed/src/mwc-top-app-bar-fixed-base.ts
